+++ lib/shared/types.ts
@@ -98,6 +98,8 @@ export interface BuildOptions extends CommonOptions {
   outbase?: string;
   /** Documentation: https://esbuild.github.io/api/#external */
   external?: string[];
+  /** Documentation: https://esbuild.github.io/api/#alias */
+  alias?: Record<string, string>;
   /** Documentation: https://esbuild.github.io/api/#loader */
   loader?: { [ext: string]: Loader };
   /** Documentation: https://esbuild.github.io/api/#resolve-extensions */